+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # 更新日志
+## v0.8.9 - 2025-12-2
+- 修复 Image/Latent/Text Switch Dual Mode节点注释栏的无故隐藏问题
+- 新增 TextSwitchDualMode 节点：文本切换器（双模式）
+<br>
+
+- fix the issue that the comment area of Image/Latent/Text Switch Dual Mode node is hidden unexpectedly
+- add the TextSwitchDualMode node: text switcher (dual mode)
+
+
 ## v0.8.8 - 2025-12-1
 - 新增PromptCardSelector提示词卡选择器节点（内置卡池）
 - Qwen3-VL API节点增加“移除思考标签”按钮
@@ -16,6 +25,7 @@ The number of Qwen3-VL API nodes has increased
 - 更新本地化中文汉化文件
 - 新增 TextModifier 节点的清理功能按钮：移除空行、空格和换行符
 <br>
+
 - Fixed the issue of the download and storage paths for the advanced and basic models of Qwen3VL
 - Fixed the input bugs of image, text, and latent space switcher ports, and added a port connection reminder form.
 - Update the Readme document